feat: 300ave agent
agent modified with learning rate appropriate to acheive 300 reward.
Can continue to iterate to produce better resutls. The 2 pickle
file commited contains the table state of the agent.

Note: Do not overwrite this pickle file as it will be used for the base
agent that can produce 300reward. Overwrite only if agent proven
to produce better results.